Initial structure rework for Form Guides. Still doesn't compile since there are a few issues to resolve yet with input-specific values.
