+++ Skylight/Skylight-API/WebModels/FormGuide.cs
@@ -1,0 +1,8 @@
+﻿namespace Skylight.WebModels
+{
+    /// <inheritdoc cref="Forms.FormGuide"/>
+    public class FormGuide
+    {
+        // TODO
+    }
+}
