+++ Skylight/Skylight-Core/Forms/IWeatherEventFormDirector.cs
@@ -1,0 +1,11 @@
+﻿using Skylight.Models;
+
+namespace Skylight.Forms
+{
+    /// <summary>
+    /// Determines how the controls on a <see cref="WeatherEvent"/> form should operate to ensure valid data.
+    /// </summary>
+    public interface IWeatherEventFormDirector : IFormDirector<WeatherEvent>
+    {
+    }
+}
